+++ CHANGELOG.md
@@ -42,6 +42,7 @@
  - fix Firefox glitch when flipping a card #255
  - remove padding hacks to mitigate Firefox bugs... #255 
  - use jQuery UI Calendar by default on all browsers to avoid issues with date format on Chrome
+ - fix #507 - Google Autocomplete doesn't works when user chose a place & then change to a custom one
 
 ## 0.1.0 (2017-12-13)
  - production deployment